Feat/#69 식사 체크리스트 레이아웃 1-1 수정 및 kt 코드 이동(팩토리명수정)
