+++ app/src/main/java/com/epi/epilog/meal/ActivityMealManageWeight.kt
@@ -1,9 +1,9 @@
-package com.epi.epilog
+package com.epi.epilog.meal
 
-import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
+import com.epi.epilog.R
 
 class ActivityMealManageWeight : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
