+++ app/src/main/res/layout/fragment_meal_checklist.xml
@@ -54,8 +54,8 @@
             android:layout_marginStart="10dp"
             android:layout_marginEnd="10dp"
             app:cv_dayViewResource="@layout/week_calendar_day_layout"
-            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/titlesContainer" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
@@ -233,9 +233,10 @@
                     android:id="@+id/meal_carbohydrate_line"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_marginStart="20dp"
+                    android:layout_marginStart="10dp"
                     android:src="@drawable/meal_today_cal_box_line"
                     app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toStartOf="@id/meal_protein_graph_container"
                     app:layout_constraintStart_toEndOf="@id/meal_carbohydrate_graph_container"
                     app:layout_constraintTop_toTopOf="parent" />
 
@@ -281,6 +282,7 @@
                     android:layout_marginStart="10dp"
                     android:src="@drawable/meal_today_cal_box_line"
                     app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toStartOf="@id/meal_fat_graph_container"
                     app:layout_constraintStart_toEndOf="@id/meal_protein_graph_container"
                     app:layout_constraintTop_toTopOf="parent" />
 
@@ -290,7 +292,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginTop="10dp"
-                    android:layout_marginEnd="20dp"
+                    android:layout_marginEnd="10dp"
                     android:orientation="vertical"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toEndOf="@id/meal_protein_line"
@@ -363,7 +365,7 @@
                     android:layout_width="334dp"
                     android:layout_height="92dp"
                     android:layout_marginStart="8dp"
-                    android:layout_marginTop="20px"
+                    android:layout_marginTop="5dp"
                     android:layout_marginEnd="8dp"
                     android:background="@drawable/medicine_redbox"
                     android:gravity="center_vertical"
@@ -415,6 +417,7 @@
                     android:background="@drawable/red_badge"
                     android:padding="5px"
                     android:text="식사시간 지남"
+                    android:visibility="visible"
                     android:textAlignment="center"
                     android:textColor="#fff"
                     android:textSize="12dp" />
@@ -424,7 +427,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="92dp"
                 android:layout_marginStart="15dp"
-                android:layout_marginTop="20px"
+                android:layout_marginTop="10dp"
                 android:layout_marginEnd="15dp"
                 android:background="@drawable/medicine_purplebox"
                 android:gravity="center_vertical"
@@ -469,16 +472,16 @@
                 <CheckBox
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-
-                    android:background="@drawable/meal_checklist_checkbtn"/>
+                    android:background="@drawable/meal_checklist_checkbtn"
+                    android:button="@null" />
             </LinearLayout>
 
             <LinearLayout
-                android:layout_width="334dp"
+                android:layout_width="330dp"
                 android:layout_height="92dp"
                 android:layout_marginStart="15dp"
-                android:layout_marginTop="20px"
                 android:layout_marginEnd="15dp"
+                android:layout_marginTop="15dp"
                 android:background="@drawable/meal_checklist_see_meal_btn"
                 android:gravity="center_vertical"
                 android:orientation="horizontal"
@@ -518,7 +521,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="92dp"
                 android:layout_marginStart="15dp"
-                android:layout_marginTop="20px"
+                android:layout_marginTop="15dp"
                 android:layout_marginEnd="15dp"
                 android:background="@drawable/medicine_orangebox"
                 android:gravity="center_vertical"
